Extract travel components and format helpers
