+++ portfolio/components/travel/TravelEmptyState.vue
@@ -1,0 +1,7 @@
+<template>
+  <div
+    class="rounded-2xl border-b border-t border-white/10 bg-white/5 p-4 text-white/70 backdrop-blur-xl"
+  >
+    No posts yet.
+  </div>
+</template>
