+++ portfolio/pages/index.vue
@@ -5,18 +5,7 @@
     </div>
     <div class="container mx-auto mt-8">
       <div class="px-3 md:px-5">
-        <h1
-          class="bg-gradient-to-b from-white via-slate-100 to-slate-300 bg-clip-text text-2xl font-semibold text-transparent"
-        >
-          Travel
-        </h1>
-        <h2 class="mt-3 text-lg leading-snug text-slate-200/90">
-          <span
-            class="bg-gradient-to-b from-white to-slate-300 bg-clip-text text-transparent"
-          >
-            Highlights
-          </span>
-        </h2>
+        <PageHeading title="Travel" subtitle="Highlights" />
       </div>
       <div id="travel-grid-card" class="mt-3 px-1 md:px-2">
         <TravelCardGrid :cards="travelCardGrid" :display-view-all="true" />
@@ -26,6 +15,10 @@
 </template>
 
 <script setup lang="ts">
+import HomeHero from "~/components/home/HomeHero.vue";
+import TravelCardGrid from "~/components/travel/TravelCardGrid.vue";
+import PageHeading from "~/components/ui/PageHeading.vue";
+
 const route = useRoute();
 
 const { data: travelCardGrid } = await useAsyncData(route.path, () => {
